+++ packages/client/hmi-client/src/components/widgets/tera-import-github-file.vue
@@ -190,7 +190,7 @@ import Dropdown from 'primevue/dropdown';
 import Breadcrumb from 'primevue/breadcrumb';
 import { createNewDatasetFromGithubFile } from '@/services/dataset';
 import { createNewArtifactFromGithubFile } from '@/services/artifact';
-import { extractPDF } from '@/services/models/extractions';
+import { extractPDF } from '@/services/extractions';
 import useAuthStore from '@/stores/auth';
 import { uploadCodeToProjectFromGithub } from '@/services/code';
